requirements.txt
flask
pillow
fpdf2
pypdfium2